api: change double to float everywhere
Warning: the commit is API/ABI backwards incompatible

The most significant change happens in sum primitive
C API, which previously took scales as an double array.
Now it takes an array of floats. For C++ we preserve
both API and ABI, thanks to overloading.

All the other changes should happen seamless...

+++ examples/simple_net.cpp
@@ -100,7 +100,7 @@ void simple_net(){
     /* AlexNet: relu
      * {batch, 96, 55, 55} -> {batch, 96, 55, 55}
      */
-    const double negative_slope = 1.0;
+    const float negative_slope = 1.0;
 
     auto relu_dst_memory = memory(conv_prim_desc.dst_primitive_desc());
 
@@ -121,9 +121,9 @@ void simple_net(){
      * beta: 0.75
      */
     const uint32_t local_size = 5;
-    const double alpha = 0.0001;
-    const double beta = 0.75;
-    const double k = 1.0;
+    const float alpha = 0.0001;
+    const float beta = 0.75;
+    const float k = 1.0;
 
     auto lrn_dst_memory = memory(relu_dst_memory.get_primitive_desc());
