+++ src/cpu/jit_primitive_conf.hpp
@@ -52,7 +52,7 @@ struct jit_conv_conf_t {
     int dilate_h, dilate_w;
     memory_format_t src_fmt;
     bool with_bias, with_relu;
-    double relu_negative_slope;
+    float relu_negative_slope;
 
     int ihp, iwp, ohp, owp;
     int nb_ic, ic_block;
@@ -147,7 +147,7 @@ struct jit_1x1_conv_conf_t {
     int stride_h, stride_w;
     memory_format_t src_fmt;
     bool with_bias, with_relu;
-    double relu_negative_slope;
+    float relu_negative_slope;
 
     int is, os;
     int ic_block, oc_block;
@@ -191,7 +191,7 @@ struct jit_gemm_conv_conf_t {
     int dilate_h, dilate_w;
     memory_format_t src_fmt;
     bool with_bias, with_relu;
-    double relu_negative_slope;
+    float relu_negative_slope;
 
     int is, os, ks;
     int ic_block, oc_block;
