+++ src/cpu/jit_sse42_conv_kernel_f32.cpp
@@ -371,7 +371,7 @@ void jit_sse42_conv_fwd_kernel_f32::generate()
 status_t jit_sse42_conv_fwd_kernel_f32::init_conf(jit_conv_conf_t &jcp,
         const convolution_desc_t &cd, const memory_desc_wrapper &src_d,
         const memory_desc_wrapper &weights_d, const memory_desc_wrapper &dst_d,
-        bool with_relu, double relu_negative_slope)
+        bool with_relu, float relu_negative_slope)
 {
     if (!mayiuse(sse42)) return status::unimplemented;
