+++ src/cpu/ref_batch_normalization.cpp
@@ -54,7 +54,7 @@ void ref_batch_normalization_fwd_t<data_type>::execute_forward() {
     const int H = conf_.H();
     const int W = conf_.W();
 
-    const double eps = conf_.desc()->batch_norm_epsilon;
+    const float eps = conf_.desc()->batch_norm_epsilon;
     const bool use_scaleshift = conf_.use_scaleshift();;
     const bool save_stats = conf_.is_training();
     const bool calculate_stats = !conf_.stats_is_src();
@@ -134,7 +134,7 @@ void ref_batch_normalization_bwd_t<data_type>::execute_backward() {
     const int H = conf_.H();
     const int W = conf_.W();
 
-    const double eps = conf_.desc()->batch_norm_epsilon;
+    const float eps = conf_.desc()->batch_norm_epsilon;
     const bool use_scaleshift = conf_.use_scaleshift();
     const bool calculate_diff_stats = !conf_.omit_stats();
